Rename package to avoid conflict with old extension
