+++ src/pt/slimereadunoriginal/build.gradle.kts
