+++ src/pt/slimereadunoriginal/res/mipmap-mdpi/ic_launcher.png
